refactor(repo): split main screen and patch research vite
Extract the Android host app home UI out of MainActivity into MainScreenContent and record the completed ERICK-141 progress.

Also add concrete follow-on AI-first Jira docs and update the nested research Vite app to a non-vulnerable dependency set with a clean npm audit.
