+++ docs/documentation/Research/vatsal/layout_design/layout_react_app/package.json
@@ -11,7 +11,7 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^4.2.1",
-    "vite": "^5.1.0"
+    "@vitejs/plugin-react": "^6.0.1",
+    "vite": "^8.0.9"
   }
 }
